+++ src/main/java/io/github/astrapi69/tree/BaseTreeNode.java
@@ -29,7 +29,6 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
 
-import io.github.astrapi69.tree.api.ITreeNode;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;