+++ CHANGELOG.md
@@ -4,12 +4,17 @@
 Version 5.5-SNAPSHOT
 -------------
 
+ADDED:
+
+- new generic id field in class SimpleTreeNode
+
 CHANGED:
 
 - the id field from the BaseTreeNode class has now a generic type
 - moved dependency data-api to a test-dependency
 - moved dependency id-generate to a test-dependency
 - method isNode in the BaseTreeNode class is now public
+- rename of class ParentIdTreeNode to TreeIdNode
 
 Version 5.4
 -------------